fixed circular import by adding new utils dir
