+++ mdagent/mainagent/__init__.py
@@ -1,4 +1,3 @@
 from .agent import MDAgent
-from .llm import _make_llm
 
-__all__ = ["MDAgent", "_make_llm"]
+__all__ = ["MDAgent"]
