+++ mdagent/subagents/agents/action.py
@@ -13,15 +13,14 @@ from langchain.prompts.chat import (
     SystemMessagePromptTemplate,
 )
 
-from ...mainagent import _make_llm
-from ...tools import PathRegistry
-from ..prompts import (
+from mdagent.subagents.prompts import (
     action_format,
     action_prefix,
     action_prefix_1,
     action_prompt,
     action_prompt_1,
 )
+from mdagent.utils import PathRegistry, _make_llm
 
 load_dotenv()
