+++ mdagent/tools/__init__.py
@@ -5,7 +5,6 @@ from .base_tools import (
     ListRegistryPaths,
     MapPath2Name,
     Name2PDBTool,
-    PathRegistry,
     PlanBVisualizationTool,
     RemoveWaterCleaningTool,
     Scholar2ResultLLM,
@@ -27,7 +26,6 @@ __all__ = [
     "SimulationFunctions",
     "VisualizationToolRender",
     "ListRegistryPaths",
-    "PathRegistry",
     "MapPath2Name",
     "Name2PDBTool",
     "get_pdb",
