+++ mdagent/tools/base_tools/__init__.py
@@ -6,7 +6,7 @@ from .clean_tools import (
 )
 from .md_util_tools import Name2PDBTool, get_pdb
 from .plot_tools import SimulationOutputFigures
-from .registry_tools import ListRegistryPaths, MapPath2Name, PathRegistry
+from .registry_tools import ListRegistryPaths, MapPath2Name
 from .search_tools import Scholar2ResultLLM
 from .setup_and_run import SetUpAndRunTool, SimulationFunctions
 from .vis_tools import (
@@ -25,7 +25,6 @@ __all__ = [
     "CheckDirectoryFiles",
     "PlanBVisualizationTool",
     "ListRegistryPaths",
-    "PathRegistry",
     "MapPath2Name",
     "Name2PDBTool",
     "get_pdb",
