+++ mdagent/tools/base_tools/clean_tools.py
@@ -5,7 +5,7 @@ from langchain.tools import BaseTool
 from openmm.app import PDBFile, PDBxFile
 from pdbfixer import PDBFixer
 
-from .registry_tools import PathRegistry
+from mdagent.utils import PathRegistry
 
 
 class CleaningTools:
