+++ mdagent/tools/base_tools/md_util_tools.py
@@ -3,7 +3,7 @@ from typing import Optional
 import requests
 from langchain.tools import BaseTool
 
-from .registry_tools import PathRegistry
+from mdagent.utils import PathRegistry
 
 
 def get_pdb(query_string, PathRegistry):
