+++ mdagent/tools/base_tools/vis_tools.py
@@ -5,7 +5,7 @@ from typing import Optional
 import nbformat as nbf
 from langchain.tools import BaseTool
 
-from .registry_tools import PathRegistry
+from mdagent.utils import PathRegistry
 
 
 class VisFunctions:
